scripts/requirements.txt
requests==2.31.0
Pillow==10.3.0
numpy==1.26.4
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
moviepy==1.0.3
scipy==1.13.0
noise==1.2.2
